RIUI-1799 final solution
