+++ src/app/domain/components/questions/create/create.component.ts
@@ -68,7 +68,9 @@ export class CreateQuestionsComponent implements OnInit {
           if(values.question.match(/[<>]/g) === null) {
             values.question = "<p>"+values.question+"</p>"
           }
-            this.form.controls.question.setValue(values.question.trim());
+          
+          values.question = this.questionService.validateHTML(values.question);
+          this.form.controls.question.setValue(values.question.trim());
         }
         values.rounds = this.roundNumber;
