+++ src/app/domain/services/question.service.spec.ts
@@ -1,11 +1,11 @@
 import { TestBed, inject } from '@angular/core/testing';
 import { QuestionService } from './question.service';
-import {HttpClientModule} from '@angular/common/http';
-import {BrowserTransferStateModule, TransferState} from '@angular/platform-browser';
-import {HttpClientTestingModule, HttpTestingController} from '@angular/common/http/testing';
-import {RouterTestingModule} from '@angular/router/testing';
-import {configMock, MockConfigService, MockTransferStateService} from './mock/config.env.mock';
-import {ConfigService} from '../../config.service';
+import { HttpClientModule } from '@angular/common/http';
+import { BrowserTransferStateModule, TransferState } from '@angular/platform-browser';
+import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
+import { RouterTestingModule } from '@angular/router/testing';
+import { configMock, MockConfigService, MockTransferStateService } from './mock/config.env.mock';
+import { ConfigService } from '../../config.service';
 
 
 describe('QuestionService', () => {
@@ -24,12 +24,12 @@ describe('QuestionService', () => {
                     provide: ConfigService,
                     useValue: configMock
                 },
-                { provide: TransferState, useClass : MockTransferStateService},
+                { provide: TransferState, useClass: MockTransferStateService },
                 // { provide: ConfigService, useClass : MockConfigService},
             ]
         });
         httpMock = TestBed.get(HttpTestingController);
-        questionService  = TestBed.get(QuestionService);
+        questionService = TestBed.get(QuestionService);
     });
 
     it('should be created', inject([QuestionService], (service: QuestionService) => {
@@ -47,32 +47,39 @@ describe('QuestionService', () => {
         expect(service.fetch(caseId, questionId)).toBeTruthy();
         expect(service.fetchAll(item)).toBeTruthy();
         expect(service.create(caseId, question)).toBeTruthy();
-    expect(service.update(caseId, questionId, question)).toBeTruthy();
-    expect(service.remove(caseId, questionId)).toBeTruthy();
-    expect(service.sendQuestions(caseId, roundId)).toBeTruthy();
-    expect(service.fetchRound(caseId, roundId)).toBeTruthy();
-}));
+        expect(service.update(caseId, questionId, question)).toBeTruthy();
+        expect(service.remove(caseId, questionId)).toBeTruthy();
+        expect(service.sendQuestions(caseId, roundId)).toBeTruthy();
+        expect(service.fetchRound(caseId, roundId)).toBeTruthy();
+    }));
 
-it('should method contain create var', inject([QuestionService], () => {
+    it('should parse string into valid HTML', inject([QuestionService], (
+        service: QuestionService
+    ) => {
+        const html = '<p>a</p><p><br></p>';
+        expect(service.validateHTML(html)).toBe('<p>a</p><p><br /></p>');
+    }));
 
-    const mockDummyData = [{ id: 1 }, { id: 2 }];
-    const caseId = '123';
-    const body = '123';
+    it('should method contain create var', inject([QuestionService], () => {
 
-    questionService.create(caseId, body).subscribe(data => {
-        expect(data).toEqual(mockDummyData);
-    });
+        const mockDummyData = [{ id: 1 }, { id: 2 }];
+        const caseId = '123';
+        const body = '123';
 
-    const mockReq = httpMock.expectOne(`${configMock.config.api_base_url}/api/caseQ/${caseId}/questions`);
-    expect(mockReq.request.method).toBe('POST');
-    expect(mockReq.request.responseType).toEqual('json');
-    expect(mockReq.request.body).toBe(body);
-    expect(mockReq.request.url).toBe( `/api/caseQ/${caseId}/questions`);
+        questionService.create(caseId, body).subscribe(data => {
+            expect(data).toEqual(mockDummyData);
+        });
+
+        const mockReq = httpMock.expectOne(`${configMock.config.api_base_url}/api/caseQ/${caseId}/questions`);
+        expect(mockReq.request.method).toBe('POST');
+        expect(mockReq.request.responseType).toEqual('json');
+        expect(mockReq.request.body).toBe(body);
+        expect(mockReq.request.url).toBe(`/api/caseQ/${caseId}/questions`);
 
 
-    mockReq.flush(mockDummyData);
-    httpMock.verify();
+        mockReq.flush(mockDummyData);
+        httpMock.verify();
 
-}));
+    }));
 
 });
