+++ src/app/domain/services/question.service.ts
@@ -58,4 +58,14 @@ export class QuestionService {
         const xUrl = `${this.configService.config.api_base_url}/api/caseQ/${caseId}/rounds/${roundId}`;
         return this.http.get(xUrl);
     }
+
+    validateHTML(html) {
+        const doc = new DOMParser().parseFromString(html, 'text/html');
+        const xmlString = new XMLSerializer().serializeToString(doc);
+        const bodyOpening = xmlString.indexOf('<body');
+        const bodyStart = bodyOpening + xmlString.substr(bodyOpening).indexOf('>') + 1;
+        const bodyEnd = xmlString.indexOf('</body>');
+        const result = xmlString.substring(bodyStart, bodyEnd);
+        return result;
+    }
 }
